pyproject.toml
[tool.poetry]
name = "Nova-Bot"
version = "0.0.1"
description = "Nova-Bot based on Nonebot2"
authors = ["Nova-Noir"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = ">=2.0.0b4"
pydantic = "1.9.1"
pymongo = "4.2.0"
nonebot-adapter-onebot = ">=2.1.1"
playwright = "^1.24.1"
aiohttp = "^3.8.1"
nonebot-plugin-apscheduler = "^0.1.4"
pycryptodome = "^3.15.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["novabot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
